Renaming application, new methods, simplify main method, formating Přejmenování aplikace, nové metody, zjednodušení main metody, přeformátování
